refactor(website): extract the desktop nav row into its own component
