+++ apps/website/src/components/shared/Nav.tsx
@@ -14,12 +14,7 @@ import { GitHubIcon } from '../ui/GitHubIcon';
 import { GITHUB_REPO_URL } from '../../lib/positioning';
 import { DEMOS, demoCtaSuffix } from '../../lib/demos';
 import { DocsContextContent } from '../docs/DocsControlPlane';
-
-const links = [
-  { label: 'Pilot to Prod', href: '/pilot-to-prod', external: false },
-  { label: 'Docs', href: '/docs', external: false },
-  { label: 'Pricing', href: '/pricing', external: false },
-];
+import { NavDesktop, links, trackNavLink } from './NavDesktop';
 
 const toAnalyticsLibrary = (library: LibraryId | null): AnalyticsLibrary => {
   switch (library) {
@@ -63,57 +58,6 @@ function CloseIcon() {
     >
       <path d="M5 5l10 10M15 5L5 15" />
     </svg>
-  );
-}
-
-function DemoDropdown() {
-  const [open, setOpen] = useState(false);
-  const ref = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const handler = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node))
-        setOpen(false);
-    };
-    document.addEventListener('mousedown', handler);
-    return () => document.removeEventListener('mousedown', handler);
-  }, []);
-  return (
-    <div ref={ref} className="nav-demo-dropdown">
-      <button
-        onClick={() => setOpen((o) => !o)}
-        className="text-sm font-mono transition-colors nav-demo-trigger"
-        aria-haspopup="true"
-        aria-expanded={open}
-      >
-        Demo{' '}
-        <span className="nav-demo-caret" data-open={open || undefined}>
-          &#9662;
-        </span>
-      </button>
-      {open && (
-        <div className="nav-demo-menu">
-          {DEMOS.map((demo) => (
-            <a
-              key={demo.key}
-              href={demo.href}
-              target="_blank"
-              rel="noopener noreferrer"
-              onClick={() => {
-                setOpen(false);
-                trackExternalLinkClick(demo.href, {
-                  surface: 'nav',
-                  cta_id: `nav_demo_${demoCtaSuffix(demo.key)}`,
-                  cta_text: demo.label,
-                });
-              }}
-              className="text-sm font-mono nav-demo-item"
-            >
-              {demo.label}
-            </a>
-          ))}
-        </div>
-      )}
-    </div>
   );
 }
 
@@ -244,30 +188,6 @@ export function Nav() {
     return () => document.removeEventListener('keydown', onKeyDown);
   }, [closeMobileMenu, open]);
 
-  const trackNavLink = (
-    label: string,
-    href: string,
-    external: boolean,
-    surface: 'nav' | 'mobile_nav'
-  ) => {
-    const slug = label
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, '_')
-      .replace(/^_|_$/g, '');
-    const ctaId: `nav_${string}` | `mobile_nav_${string}` =
-      surface === 'nav' ? `nav_${slug}` : `mobile_nav_${slug}`;
-    if (external) {
-      trackExternalLinkClick(href, { surface, cta_id: ctaId, cta_text: label });
-      return;
-    }
-    trackCtaClick({
-      surface,
-      destination_url: href,
-      cta_id: ctaId,
-      cta_text: label,
-    });
-  };
-
   return (
     <>
       <nav
@@ -282,66 +202,7 @@ export function Nav() {
           </Link>
 
           {/* Desktop links */}
-          <div className="hidden lg:flex items-center gap-8">
-            {links.map((l) =>
-              l.external ? (
-                <a
-                  key={l.href}
-                  href={l.href}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  onClick={() => trackNavLink(l.label, l.href, true, 'nav')}
-                  className="text-sm font-mono transition-colors nav-link"
-                >
-                  {l.label}
-                </a>
-              ) : (
-                <Link
-                  key={l.href}
-                  href={l.href}
-                  onClick={() => trackNavLink(l.label, l.href, false, 'nav')}
-                  className="text-sm font-mono transition-colors nav-link"
-                >
-                  {l.label}
-                </Link>
-              )
-            )}
-            <DemoDropdown />
-            <a
-              href={GITHUB_REPO_URL}
-              target="_blank"
-              rel="noopener noreferrer"
-              onClick={() =>
-                trackExternalLinkClick(
-                  GITHUB_REPO_URL,
-                  {
-                    surface: 'nav',
-                    cta_id: 'nav_github',
-                    cta_text: 'GitHub',
-                  }
-                )
-              }
-              className="transition-colors nav-link"
-              aria-label="GitHub repository"
-            >
-              <GitHubIcon />
-            </a>
-            <Button
-              variant="primary"
-              size="md"
-              href="/contact"
-              onClick={() =>
-                trackCtaClick({
-                  surface: 'nav',
-                  destination_url: '/contact',
-                  cta_id: 'nav_talk_to_us',
-                  cta_text: 'Talk to Us',
-                })
-              }
-            >
-              Talk to Us
-            </Button>
-          </div>
+          <NavDesktop />
 
           {/* Mobile hamburger */}
           <button
